Rename file name to invoices.service.ts
